pyproject.toml
[project]
name = "ai-agent-study"
version = "0.1.0"
description = "Example of using LangGraph for multi-agent workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.0.4",
    "feedparser>=6.0.11",
    "grandalf>=0.8",
    "graphviz>=0.21",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.9",
    "lxml[html-clean]>=5.4.0",
    "newspaper3k>=0.2.8",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pre-commit>=4.2.0",
    "pytest>=8.0.0",
]

[tool.ruff]
# 기본 설정
line-length = 119
target-version = "py312"

[tool.ruff.format]
# Black 호환 포맷팅
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
# 기본 규칙
select = ["E", "F", "W", "B", "I"]

# notebook 파일에서 무시할 규칙들
[tool.ruff.lint.per-file-ignores]
"*.ipynb" = [
    "E402",  # Module level import not at top of file (notebook cells are independent)
    "F401",  # Unused imports (common in notebooks for exploration)
    "E703",  # Statement ends with a semicolon (notebook display suppression)
]
